requirements.txt
asgiref==3.9.2
crispy-bootstrap5==2025.6
Django==5.2.7
django-crispy-forms==2.4
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
